refactor: separate ffis from code refactor: rename libraries
