+++ crates/batcher/go_verifiers_lib/verifier.go
