+++ crates/batcher/src/circom/verifier.rs
@@ -1,31 +1,5 @@
 use aligned_sdk::common::types::ProvingSystemId;
-
-#[derive(Copy, Clone, Debug)]
-#[repr(C)]
-pub struct ListRef {
-    data: *const u8,
-    len: usize,
-}
-
-impl From<Vec<u8>> for ListRef {
-    fn from(v: Vec<u8>) -> Self {
-        Self::from(v.as_slice())
-    }
-}
-
-impl From<&Vec<u8>> for ListRef {
-    fn from(v: &Vec<u8>) -> Self {
-        Self::from(v.as_slice())
-    }
-}
-
-impl From<&[u8]> for ListRef {
-    fn from(v: &[u8]) -> Self {
-        let len = v.len();
-        let data = v.as_ptr().cast();
-        ListRef { data, len }
-    }
-}
+use crate::ffi::circom_ffi::VerifyCircomGroth16ProofBN128;
 
 pub fn verify_circom(
     proving_system: &ProvingSystemId,
@@ -43,12 +17,4 @@ pub fn verify_circom(
         },
         _ => false,
     }
-}
-
-extern "C" {
-    pub fn VerifyCircomGroth16ProofBN128(
-        proof: ListRef,
-        public_input: ListRef,
-        verification_key: ListRef,
-    ) -> bool;
 }
