+++ crates/batcher/src/ffi/mod.rs
@@ -1,0 +1,3 @@
+mod list_ref;
+pub mod circom_ffi;
+pub mod gnark_ffi;
